test: parallelize setup cases
